Improve flat reader
